supernova-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
argon2-cffi==23.1.0
    # via -r requirements.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
bidict==0.23.1
    # via -r requirements.in
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.1.8
    # via uvicorn
cryptography==44.0.0
    # via
    #   pyspnego
    #   smbprotocol
fastapi==0.115.8
    # via
    #   -r requirements.in
    #   fastapi-restful
fastapi-restful==0.6.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
lxml==5.3.1
    # via -r requirements.in
multidict==6.1.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via typing-inspect
numpy==2.2.2
    # via
    #   -r requirements.in
    #   pandas
orjson==3.10.15
    # via -r requirements.in
packaging==24.2
    # via gunicorn
pandas==2.2.3
    # via -r requirements.in
psutil==5.9.8
    # via fastapi-restful
pyasn1==0.6.1
    # via rsa
pycparser==2.22
    # via cffi
pycryptodome==3.21.0
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   fastapi
    #   fastapi-restful
pydantic-core==2.27.2
    # via pydantic
pymupdf==1.25.3
    # via -r requirements.in
pyodbc==5.2.0
    # via -r requirements.in
pyspnego==0.11.2
    # via smbprotocol
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.1
    # via pandas
rsa==4.9
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smbprotocol==1.15.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via -r requirements.in
starlette==0.45.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via -r requirements.in
tzdata==2025.1
    # via pandas
uvicorn==0.34.0
    # via -r requirements.in
xmltodict==0.14.2
    # via -r requirements.in
